Fix issues with dependency installation and build.
The package name in glide.yaml did not conform to Go packing naming
conventions, causing glide to be unable to properly determine what
package it was working in.

Some imports of packages did not use the standard "site/user/package"
naming convention, which confused both go and glide.

The "go install" lines in the Makefile were removed as they were causing
errors.